Add JWT authentication middleware and role-based authorization

+++ controllers/problemController.js
@@ -1,6 +1,7 @@
 const Problem = require("../models/Problem");
 const asyncHandler = require("../utils/asyncHandler");
 const ApiError = require("../utils/ApiError");
+const ApiResponse = require("../utils/apiResponse");
 
 const createProblem = asyncHandler(async (req, res) => {
   const { title, description, difficulty, sampleInput, sampleOutput } =
@@ -21,36 +22,41 @@ const createProblem = asyncHandler(async (req, res) => {
     sampleInput,
     sampleOutput,
   });
-  return res.status(201).json({
-    success: true,
-    message: "Problem created successfully.",
-    problem,
-  });
+  return res
+    .status(201)
+    .json(new ApiResponse(201, problem, "Problem created successfully."));
 });
 
 const getProblems = asyncHandler(async (req, res) => {
   const problems = await Problem.find();
-  return res.status(200).json({
-    success: true,
-    problems,
-  });
+  return res
+    .status(200)
+    .json(new ApiResponse(200, problems, "Problems fetched successfully."));
 });
 
 const getProblemById = asyncHandler(async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new ApiError(400, "Invalid problem ID");
+  }
+
   const problem = await Problem.findById(id);
 
   if (!problem) {
     throw new ApiError(404, "Problem not found");
   }
-  return res.status(200).json({
-    success: true,
-    problem,
-  });
+  return res
+    .status(200)
+    .json(new ApiResponse(200, problem, "Problem fetched successfully."));
 });
 
 const updateProblem = asyncHandler(async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new ApiError(400, "Invalid problem ID");
+  }
 
   const updatedData = req.body;
 
@@ -62,22 +68,25 @@ const updateProblem = asyncHandler(async (req, res) => {
   if (!problem) {
     throw new ApiError(404, "Problem not found.");
   }
-  return res.status(200).json({
-    success: true,
-    problem,
-  });
+  return res
+    .status(200)
+    .json(new ApiResponse(200, problem, "Problem updated successfully."));
 });
 
 const deleteProblem = asyncHandler(async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new ApiError(400, "Invalid problem ID");
+  }
+
   const problem = await Problem.findByIdAndDelete(id);
   if (!problem) {
     throw new ApiError(404, "Problem does not exist.");
   }
-  return res.status(200).json({
-    success: true,
-    message: "Problem deleted successfully.",
-  });
+  return res
+    .status(200)
+    .json(new ApiResponse(200, problem, "Problem deleted successfully."));
 });
 
 module.exports = {
